requirements.txt
azure==0.8.4
beautifulsoup4
wheezy.template
pytz